# Routes/api-docs-rev1.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443/
  description: Generated server url
paths:
  /wsRuta/ruta/estatus:
    put:
      tags:
      - rest-controller-route
      operationId: statusRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request'
        required: true
      responses:
        500:
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                - type: object
                - $ref: '#/components/schemas/Response'
        206:
          description: Partial Content
          content:
            '*/*':
              schema:
                type: object
        204:
          description: No Content
          content:
            '*/*':
              schema:
                type: object
        405:
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: object
        200:
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Response'
  /wsRuta/ruta/creacion:
    post:
      tags:
      - rest-controller-route
      operationId: createRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request'
        required: true
      responses:
        500:
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                - type: object
                - $ref: '#/components/schemas/Response'
        206:
          description: Partial Content
          content:
            '*/*':
              schema:
                type: object
        204:
          description: No Content
          content:
            '*/*':
              schema:
                type: object
        405:
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: object
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /wsRuta/ruta/cancelar:
    post:
      tags:
      - rest-controller-route
      operationId: cancelRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request'
        required: true
      responses:
        500:
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                - type: object
                - $ref: '#/components/schemas/Response'
        206:
          description: Partial Content
          content:
            '*/*':
              schema:
                type: object
        204:
          description: No Content
          content:
            '*/*':
              schema:
                type: object
        405:
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: object
        200:
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Response'
  /wsRuta:
    get:
      tags:
      - rest-controller-route
      operationId: index
      responses:
        500:
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                - type: object
                - $ref: '#/components/schemas/Response'
        206:
          description: Partial Content
          content:
            '*/*':
              schema:
                type: object
        204:
          description: No Content
          content:
            '*/*':
              schema:
                type: object
        405:
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: object
        200:
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Response'
  /wsRuta/:
    get:
      tags:
      - rest-controller-route
      operationId: index_1
      responses:
        500:
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                - type: object
                - $ref: '#/components/schemas/Response'
        206:
          description: Partial Content
          content:
            '*/*':
              schema:
                type: object
        204:
          description: No Content
          content:
            '*/*':
              schema:
                type: object
        405:
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: object
        200:
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Response'
  /wsRuta/ruta/documento:
    get:
      tags:
      - rest-controller-route
      operationId: getDocRoute
      parameters:
      - name: req
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Request'
      responses:
        500:
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                - type: object
                - $ref: '#/components/schemas/Response'
        206:
          description: Partial Content
          content:
            '*/*':
              schema:
                type: object
        204:
          description: No Content
          content:
            '*/*':
              schema:
                type: object
        405:
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: object
        200:
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Response'
  /wsRuta/consultas/estatus:
    get:
      tags:
      - rest-controller-validation
      operationId: createRoute_1
      parameters:
      - name: req
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Request'
      responses:
        500:
          description: Internal Server Error
          content:
            '*/*':
              schema:
                oneOf:
                - type: object
                - $ref: '#/components/schemas/Response'
        206:
          description: Partial Content
          content:
            '*/*':
              schema:
                type: object
        204:
          description: No Content
          content:
            '*/*':
              schema:
                type: object
        405:
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: object
        200:
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Response'
components:
  schemas:
    Response:
      type: object
      properties:
        codigo:
          type: string
        mensaje:
          type: string
        folio:
          type: string
        resultado:
          type: object
    Request:
      required:
      - folios
      - idCentro
      type: object
      properties:
        idPeticion:
          type: integer
          format: int64
        idCentro:
          type: integer
          format: int32
        folios:
          type: array
          items:
            type: string
        idTransporte:
          type: integer
          format: int32
        idUnidad:
          type: integer
          format: int32
        largo:
          type: number
          format: double
        alto:
          type: number
          format: double
        ancho:
          type: number
          format: double
        peso:
          type: number
          format: double
        estatus:
          type: string
        fecha:
          type: string
        tipoRuta:
          type: integer
          format: int32
        motivo:
          type: integer
          format: int32
        secuencia:
          type: integer
          format: int32
        external:
          type: boolean
        estibador:
          type: string
        checador:
          type: string